Add necessary magic to get tests working in Android Studio.
This requires the Android SDK to be the *last* entry in the orderEntry
list in sdk.iml. The run configuration should be a Gradle configuration
with a 'test' task, operating on matrix-android-sdk (NOT :sdk). As this
is just effectively running 'gradle test', the test output will be in
the console window: it doesn't pretty-print JUnit test output.