Add app-global GcmRegistrationManager to register for push services
Registration must be done both with the GCM server and with the matrix
home server, supplying a token acquired from the GCM server.

Registration is done in an AsyncTask, therefore the Matrix pushers API
can be called in a synchronous fashion without impacting responsiveness.

Registration of the pusher with the Matrix server is currently done on
every startup by SplashActivity but should eventually become a one-time
background task that is not tied to a specific activity.