IMXStore::getRoomMessages becomes getEarlierMessages.
It returns a chunk of messages from the store from a start token.

The chat loading time was increased sometimes a lot when the stored room history was huge e.g when the user back paginated a lot (and got hundred messages.)

By now, the behaviour should be the same as a remote request.