Add ability to use filter from server
This adds Interfaces to create a Filter which can be send to the server.
This especially handles the DataSaveMode.
The Filter is therefore uploaded in lazy-load manner and used once the
server responds with a filter_id.

This also deprecates "setUseDataSaveMode" and now setSyncFilter should
be used. This is because the app might store the filter-id so the call
for the filter-id could be avoided if the app handles that directly

Signed-Off-by: Matthias Kesler <krombel@krombel.de>