Bug Fix - Killed Application miss the first notification
When the application is killed the first push notification is never displayed.
This was due to a bug in the RoomAccountData definition. Indeed some items of this class were not Serializable.
This make fail the loading of RoomAccountData from the store on new application launch
-> The store is considered corrupted -> the stream token is reseted
The first event which awakes the app is handled during an initial sync, so no push is applied for it...